backend/requirements.txt
asgiref
black
blinker
CacheControl
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
Deprecated
Django
django-cors-headers
djangorestframework
exceptiongroup
firebase-admin
flake8
Flask
gcloud
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
httplib2
idna
iniconfig
itsdangerous
Jinja2
jwcrypto
jws
MarkupSafe
mccabe
msgpack
mypy-extensions
numpy
oauth2client
packaging
pandas
pathspec
platformdirs
pluggy
proto-plus
protobuf
pyasn1
pyasn1-modules
pycodestyle
pycparser
pycryptodome
pyflakes
PyJWT
pyparsing
Pyrebase4
pytest
pytest-django
python-dateutil
python-decouple
python-firebase
python-jwt
pytz
requests
requests-toolbelt
rsa
scipy
six
sqlparse
tomli
typing_extensions
tzdata
uritemplate
urllib3
Werkzeug
wrapt
